requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
boto3==1.26.152
botocore==1.29.152
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
Django==4.2.3
django-appconf==1.0.5
django-celery-beat==2.5.0
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-filter==23.2
django-imagekit==4.1.0
django-jalali-date==1.0.2
django-js-asset==2.0.0
django-mptt==0.14.0
django-nine==0.2.7
django-storages==1.13.2
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
idna==3.4
inflection==0.5.1
ippanel==2.0.7
itypes==1.2.0
jdatetime==4.1.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
kombu==5.3.1
MarkupSafe==2.1.3
packaging==23.1
pilkit==2.0
Pillow==9.5.0
prompt-toolkit==3.0.39
psycopg2==2.9.6
PyJWT==2.7.0
pyrsistent==0.19.3
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.4.0
